Time: 56 ms (19.88%), Space: 61.6 MB (70.14%) - LeetHub

+++ 1552-magnetic-force-between-two-balls/1552-magnetic-force-between-two-balls.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    bool isGuessValid(vector<int>& arr , int m , long long guess){
+    bool isGuessValid(vector<int>& arr , int m , int guess){
         int ballPlaced = 1 ; 
         int pos = 0;
         for(int i = 1 ; i < arr.size() ; i++){
@@ -14,10 +14,10 @@ public:
     int maxDistance(vector<int>& position, int m) {
         int n = position.size() ;
         sort(position.begin() , position.end());
-        long long low = 1 , high = position[n-1] - position[0] ;
+        int low = 1 , high = position[n-1] - position[0] ;
         int ans = -1 ;
         while(low<=high){
-            long long guess = (low + high)/2;
+            int guess = (low + high)/2;
             if(isGuessValid(position , m , guess)){
                 ans = guess ;
                 low = guess+1 ;
@@ -27,3 +27,25 @@ public:
         return ans ;
     }
 };
+// We are binary searching the minimum distance between balls.
+//
+// For a guessed distance 'd', greedily place balls from left to right.
+//
+// Always place the next ball at the earliest position whose
+// distance from the previously placed ball is at least 'd'.
+
+// If we can place >= m balls, then distance 'd' is feasible.
+// We try to maximize it.
+
+// As distance increases, placing balls becomes harder.
+
+// Example:
+// d = 1 -> possible
+// d = 2 -> possible
+// d = 3 -> possible
+// d = 4 -> not possible
+// d = 5 -> not possible
+
+// Pattern:
+// True True True False False
+// We need the last valid distance, hence binary search.
